+++ query/src/api/http/v1/mod.rs
@@ -19,5 +19,8 @@ pub mod cluster;
 pub mod config;
 
 pub mod hello;
+#[cfg(test)]
+mod hello_test;
 pub mod logs;
+#[cfg(test)]
 mod logs_test;
